requirements.txt
google-api-python-client
httplib2
oauth2client
google-auth-oauthlib
# discord.py==1.7.3
# discord-py-slash-command==3.0.3
# discord-py-interactions==5.1.0
# discord-ui
py-cord